Custom hook working like a boss. Why was this so complicated in lecture?
