+++ client/src/hooks/useForm.js
@@ -1,1 +1,31 @@
 // write your custom hook here to control your checkout form
+
+import { useState } from "react";
+
+const initialValue = {
+  firstName: "",
+  lastName: "",
+  address: "",
+  city: "",
+  state: "",
+  zip: "",
+};
+
+const useForm = () => {
+  const [showSuccessMessage, setShowSuccessMessage] = useState(false);
+
+  const [values, setValues] = useState(initialValue);
+  const handleChanges = (e) => {
+    setValues({ ...values, [e.target.name]: e.target.value });
+  };
+
+  const handleSubmit = (e) => {
+    e.preventDefault();
+    setShowSuccessMessage(true);
+  };
+
+
+  return [values, handleChanges, handleSubmit, showSuccessMessage]
+}
+
+export default useForm
